Move config to its own module and use derive_builder
