+++ metrics/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 ddcommon = { path = "../ddcommon" }
+derive_builder = "0.12.0"
 hyper = "0.14.25"
 hyper-rustls = {version = "0.23", default-features = false, features = ["native-tokio", "http1", "tls12"]}
 serde = "1.0.159"
