+++ metrics/src/lib.rs
@@ -1,6 +1,10 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
+#[macro_use]
+extern crate derive_builder;
+
 pub mod agent;
+pub mod config;
 pub mod metrics;
 pub mod payload;
